Use long instead of uint for Unix dates to support dates before 1970.
2006-01-22  Gabriel Burt  <gabriel.burt@gmail.com>

	* src/Db.cs:
	* src/ImportStore.cs:
	* src/PhotoStore.cs: Use long instead of uint for Unix dates to
	support dates before 1970. Patch by Fredrik Noring, noring@nocrew.org,
	and updated by Curtis Hovey.

+++ src/Db.cs
@@ -285,14 +285,14 @@ public class Db : IDisposable {
 
 
 public class DbUtils {
-	public static DateTime DateTimeFromUnixTime (uint unix_time)
+	public static DateTime DateTimeFromUnixTime (long unix_time)
 	{
 		DateTime date_time = new DateTime (1970, 1, 1).ToLocalTime ();
 		return date_time.AddSeconds (unix_time);
 	}
 
-	public static uint UnixTimeFromDateTime (DateTime date_time)
+	public static long UnixTimeFromDateTime (DateTime date_time)
 	{
-		return (uint) (date_time - new DateTime (1970, 1, 1).ToLocalTime ()).TotalSeconds;
+		return (long) (date_time - new DateTime (1970, 1, 1).ToLocalTime ()).TotalSeconds;
 	}
 }
